As int

+++ purple_team/project_statuses.sql
@@ -5,6 +5,6 @@ select
   raw_score, 
   submitted_on, 
   scored_on, 
-  raw_score is null and status in ('rating','performed','scored','inRevision') as turned_in_but_not_scored
+  (raw_score is null and status in ('rating','performed','scored','inRevision'))::int as turned_in_but_not_scored
 from project_assignments as pas
 ;
